+++ .gitignore
@@ -52,3 +52,7 @@ pwa
 
 .DS_store
 /image-component.code-workspace
+/test-results/
+/playwright-report/
+/blob-report/
+/playwright/.cache/
