+++ package.json
@@ -7,7 +7,9 @@
     "ci": "webpack --config tools/webpack/config/integration",
     "dev": "webpack serve --config tools/webpack/config/dev --hot",
     "start": "npm run dev",
-    "test": "jest",
+    "test": "npm run jest & npm run playwright",
+    "jest": "jest ./__tests__/jest",
+    "playwright": "npx playwright test",
     "watch": "npm run watch-dev",
     "watch-dev": "webpack --config tools/webpack/config/integration --watch"
   },
@@ -25,14 +27,16 @@
     },
     "testEnvironment": "jsdom",
     "testEnvironmentOptions": {
-      "customExportConditions": [] 
+      "customExportConditions": []
     }
   },
   "devDependencies": {
     "@babel/core": "^7.8.0",
     "@babel/plugin-transform-react-jsx": "^7.23.4",
     "@babel/preset-env": "^7.24.5",
+    "@playwright/test": "^1.45.2",
     "@stormid/eslint-config": "^1.0.2",
+    "@types/node": "^20.14.10",
     "babel-jest": "^29.7.0",
     "babel-loader": "^9.1.3",
     "browser-sync": "^3.0.2",
